Avoid segmenation fault when accessing empty vectors
If columns in the EOP file are not populated with values, the read
routine re-uses the last available values by calling back() on a vector.
This causes a segmentation fault if the vector is empty. An exception is
now thrown in this case and the program exits properly with an error.