Improve extraction of a-priori station position
The extraction of the position a-priori values from the station position
file has been changed to search for two consecutive epochs which include
the start epoch of the processing interval and return the position
corresponding to the lower epoch. This change ensures that a position is
always found even if the processing interval is shorter than the spacing
of epochs in the station position file. For consistency with simulated
data, midnight of the start epoch is used. Warnings are added if the
station position file cannot be read or if no position is found.